second set of refactors, moved Value (will be refactored to Word in a prior MR, so may cause conflicts later) and extracted Field into the overall dataParts package
