+++ src/main/java/group1/mips_simulator/components/dataParts/Value.java
@@ -1,4 +1,4 @@
-package group1.mips_simulator.components;
+package group1.mips_simulator.components.dataParts;
 
 import group1.mips_simulator.Utility;
 import group1.mips_simulator.components.Config;
